requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
argcomplete==1.8.1
asttokens==2.2.1
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
blinker==1.4
cachetools==5.2.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
cloud-init==22.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.7
cryptography==2.8
cycler==0.11.0
dbus-python==1.2.16
decorator==5.1.1
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
executing==1.2.0
Flask==2.2.2
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
httplib2==0.14.0
hydra-core==1.3.1
hyperlink==19.0.0
idna==2.8
imantics==0.1.12
importlib-metadata==6.0.0
importlib-resources==5.10.2
incremental==16.10.1
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.0
omegaconf==2.3.0
opencv-contrib-python==4.7.0.68
opencv-python==4.3.0.36
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.4.0
progressbar==2.5
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.14.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-gnupg==0.4.5
pytz==2022.7.1
pyxdg==0.26
PyYAML==5.3.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
scipy==1.10.0
seaborn==0.12.2
SecretStorage==2.3.1
sentry-sdk==1.13.0
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
smmap==5.0.0
sos==4.3
ssh-import-id==5.10
stack-data==0.6.2
systemd-python==234
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.8.1
Twisted==18.9.0
typing-extensions==4.4.0
ubuntu-advantage-tools==27.9
Ubuntu-Make==18.9
ufw==0.36
ultralytics==8.0.11
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.6
Werkzeug==2.2.2
xmljson==0.2.1
zipp==3.11.0
zope.interface==4.7.1
